Split ModuleGenerator, NpmPackage and TypeDefinitionGenerator into seperate files
