+++ packages/generator-typescript/templates/gjs/gjs.d.ts
@@ -3,10 +3,10 @@
 <%_ if(!package){ -%>
     <%_ if(GObject){ -%>
 /// <reference path="../gobject-2.0/gobject-2.0.d.ts" />
-<% } -%>
-<%_ if(GLib){ -%>
+    <% } -%>
+    <%_ if(GLib){ -%>
 /// <reference path="../glib-2.0/glib-2.0.d.ts" />
-<% } -%>
+    <% } -%>
 /// <reference path="./gettext.d.ts" />
 /// <reference path="./system.d.ts" />
 /// <reference path="./cairo.d.ts" />
